updating the modules
